refactor: move codegen to /frontend
