+++ frontend/package.json
@@ -2,6 +2,7 @@
   "private": true,
   "scripts": {
     "codegen": "ts-node --project codegen/tsconfig.json --transpile-only codegen",
+    "generate": "graphql-codegen",
     "clean": "rimraf build coverage",
     "start": "parcel serve src/index.html",
     "build": "parcel build src/index.html",
@@ -30,6 +31,11 @@
   "devDependencies": {
     "@babel/core": "^7.7.7",
     "@babel/preset-env": "^7.9.6",
+    "@graphql-codegen/cli": "^1.13.5",
+    "@graphql-codegen/near-operation-file-preset": "^1.13.5",
+    "@graphql-codegen/typescript": "^1.13.5",
+    "@graphql-codegen/typescript-operations": "^1.13.5",
+    "@graphql-codegen/typescript-urql": "^1.13.5",
     "@types/glob": "^7.1.1",
     "@types/jest": "^25.2.1",
     "@types/react": "^16.9.34",
