--- frontend/src/build
@@ -1,1 +1,0 @@
-../../build
