+++ package.json
@@ -1,17 +1,6 @@
 {
   "private": true,
-  "scripts": {
-    "generate": "graphql-codegen"
-  },
   "dependencies": {
     "@openzeppelin/contracts": "^3.0.1"
-  },
-  "devDependencies": {
-    "@graphql-codegen/cli": "^1.13.5",
-    "@graphql-codegen/near-operation-file-preset": "^1.13.5",
-    "@graphql-codegen/typescript": "^1.13.5",
-    "@graphql-codegen/typescript-operations": "^1.13.5",
-    "@graphql-codegen/typescript-urql": "^1.13.5",
-    "graphql": "^15.0.0"
   }
 }
